moves plans helper fns to another file
